+++ src/love/config.h
@@ -19,12 +19,12 @@ namespace love {
  *
  * @code
  * def conf(t) {
- *   t.version = "0.19.0"       // Version of ChaiLove
+ *   t.version = "0.20.0"       // Version of ChaiLove
  *   t.identity = "mygame"      // Machine name of your game
  *   t.window.title = "My Game" // Human-readable name
  *   t.window.width = 1024      // Game width
  *   t.window.height = 768      // Game height
- *   t.console = false          // Developer console
+ *   t.console = false          // Developer console, toggle with tilde
  * }
  * @endcode
  */