+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to [ChaiLove](https://github.com/RobLoach/ChaiLove) will be 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.20.0 - 2018-04-22
+### Added
+- No Game demo when ChaiLove is loaded without content
+- Version mismatch warning when content's ChaiLove version doesn't match core's
+- `love.data.hash()` `sha1` support
+- `love.system.getUsername()`
+
 ## 0.19.0 - 2018-04-13
 ### Added
 - Android build